requirements.txt
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers